Create FactCode
This is the starter code for the Alexa Skill Facts. I changed a bit to make it say workouts instead of facts.
I need to make it so that the workouts aren't random. I would like it to give the user a specific workout depending on the number of days it has been since they enable the skill. 
I'm a beginner at coding, so I don't possess even the most basic concepts. I'm hoping to learn a lot from you all.